fixed channel assignment issue

+++ pwm.h
@@ -7,7 +7,7 @@
 
 #define PICO_THRUST_DEFUALT_WIDTH 1000;
 
-void thruster_setup(thruster_t *thruster, const char *id_, unsigned int gpio_);
+void thruster_setup(thruster_t *thruster, const char *id_, uint8_t gpio_);
 void thruster_init(thruster_t *thruster);
 
 void thruster_set(thruster_t *thruster, float width);
